backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
transformers>=4.44
torch>=2.0
torchvision>=0.15
torchaudio>=2.0
pandas>=2.0
matplotlib>=3.5
Pillow>=10.0
scikit-learn>=1.3
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
